replaces "flag" lib with "cobra"
it's much easier to add more flags and provides a nice, clean
separation of concerns
